requirements.txt
asgiref==3.11.0
beautifulsoup4
django-storages[s3]
boto3
crispy-bootstrap5==2025.6
diff-match-patch==20241021
dj-database-url
Django==5.2.10
django-crispy-forms==2.5
django-import-export==4.4.0
django-simple-history==3.11.0
gunicorn
pillow==12.1.0
psycopg2-binary
python-dotenv==1.2.1
requests
scikit-learn
sqlparse==0.5.5
tablib==3.9.0
typing_extensions==4.15.0
tzdata==2025.3
whitenoise
